[refactor] repository layer를 infrastructure layer 로 이동
