+++ src/main/java/junseok/snr/wallet/application/service/TransactionService.java
@@ -6,8 +6,8 @@ import junseok.snr.wallet.api.controller.dto.WithdrawDto;
 import junseok.snr.wallet.domain.Transaction;
 import junseok.snr.wallet.domain.TransactionType;
 import junseok.snr.wallet.domain.Wallet;
-import junseok.snr.wallet.api.repository.TransactionRepository;
-import junseok.snr.wallet.api.repository.WalletRepository;
+import junseok.snr.wallet.infrastructure.repository.TransactionRepository;
+import junseok.snr.wallet.infrastructure.repository.WalletRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
