+++ src/main/java/junseok/snr/wallet/schedule/TransactionScheduler.java
@@ -5,8 +5,8 @@ import junseok.snr.wallet.domain.Transaction;
 import junseok.snr.wallet.domain.TransactionStatus;
 import junseok.snr.wallet.domain.TransactionType;
 import junseok.snr.wallet.domain.Wallet;
-import junseok.snr.wallet.api.repository.TransactionRepository;
-import junseok.snr.wallet.api.repository.WalletRepository;
+import junseok.snr.wallet.infrastructure.repository.TransactionRepository;
+import junseok.snr.wallet.infrastructure.repository.WalletRepository;
 import junseok.snr.wallet.application.service.TransactionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
